Fixed race condition in state package tests.
The monitoring Goroutines here should have been waiting for a change
from the initial state of the tracker (1) rather than the 0 state. If
they invoked WaitForChange before the main Goroutine's change
notification came in, they'd see the 1 state and then receive another
notification when the tracker switched to the 2 state.