Refined symlink implementation and added Windows support.
This commit refines Mutagen's symlink synchronization support. It splits
symlink synchronization into three modes: sane (the default), ignore,
and POSIX raw. None of these are quite the same as the previous
behavior, but the default behavior brought about should be safe to
continue any sessions started under the previous version.

Sane synchronization synchronizes only those symlinks which can be
safely synchronized (i.e. relative symlinks with limited special
characters which don't escape the synchronization root). It supports
both Windows and POSIX, but on Windows obviously relies on having
permissions to create symlinks. This is left as an exercise to the user.
The inability to create a symlink does not halt synchronization.
However, if an invalid symlink is detected anywhere in the
synchronization root, it is treated as a scan error, and synchronization
is halted until the symlink is removed or fixed. The reason that we have
to do this (as opposed to just ignoring invalid ones) is that if a
symlink suddenly became valid or invalid, it would appear/disappear from
the snapshot if we just ignored invalid ones, which obviously wouldn't
work with the reconciliation algorithm. There's no way to solve that
other than blocking the whole synchronization from moving forward until
the problematic symlink is fixed.

Ignore behavior is simply to ignore all symlinks. This can be used in
the case where there's a synchronization root that contains invalid
symlinks that can't or shouldn't be removed.

POSIX raw behavior propagates symlinks in their raw form. It is only
supported on POSIX platforms, and scan will throw an error if it is used
on Windows. This behavior is fine on POSIX platforms, but for security
reasons and portability reasons, it is not recommended. It is
essentially the same as the old behavior, but if specified will now
raise an error on Windows during scan.

All together, I think these three modes give a combination of sane
behavior and control. Additional symlink integration between Windows and
POSIX may be possible, but it will be extremely difficult.