Fixed SSH transport for cmd.exe shells.
It turns out that cmd.exe shells don't like '/' as a path separator when
invoked over SSH. In fact, even invoked locally, their support for '/'
is not great. You cannot invoke a command with '/' in the path to the
command.

This commit modifies the SSH logic to detect responses from a cmd.exe
shell and retry accordingly.