Relocated POSIX command-not-found detection.
Since we do Windows command-not-found detection in the agent package,
I've relocated the corresponding POSIX code to the agent package as
well. The exact separation of logic between the ssh and agent packages
isn't super well-defined, but I think of the ssh package as holding the
stuff that doesn't manage platform probing, and I think this function is
thus better off living in the agent package.