Modified controller to wait for reconnect on synchronization error.
The controller was previously going straight to a reconnect on
synchronization errors. If reconnect is fast (e.g. to localhost), then
the daemon just spins, because the error usually keeps reoccurring. This
commit modifies the wait interval on synchronization failure to be the
same as that on connection failure. The change actually fit nicely with
the existing code.

Aside from CPU usage, this problem almost manifested itself with the
notify package. It (or FSEvents, I'm not sure which) has some internal
race condition where starting and stopping watches with high frequency
causes a segfault. It's probably relying on some sort of timer. It can
handle relatively high frequency (even > 1 Hz), but when you get into
the 5-10 Hz range it starts failing.