Modify sync.Transition to return entries.
It didn't really make sense to have sync.Transition return Change
objects. For one, we already had the transition paths on the
controller, and thus we were wasting bandwidth sending then back.
Additionally, the way sync.Transition was using Change didn't agree with
the invariants enforced by Change.EnsureValid, since both Old and New
could become nil.

This commit modifies sync.Transition (and the code that sits on top of
it) to return sync.Entry objects. They are converted to changes for the
ancestor on the controller.

One small hack was required: gob won't encode nil pointers in a slice of
pointers. Thus, we have to wrap entries in the Archive type when sending
through the remote endpoint protocol. This protocol will probably switch
over to Protocol Buffers at some point, at which point we can remove
this wrapping, but that's going to require a lot of changes, and we
might be better off just waiting until gRPC replaces this custom
protocol.