Added comment for posixCommandNotFoundExitCode.
This value (127) doesn't seem to correspond to any errno values - it
seems to be something that comes from the shell itself. Well, at least
bash and zsh. It's unclear whether or not all other shells use this
value. In any case, we have to manually define it, so it's best to add a
justification.