Switched to internal FLATE compression for agent endpoint streams.
The SSH compression was not giving us ideal results (we were about 3.6x
rsync transfer time), so we've switched to using compress/flate for
compressing agent endpoint streams. The compress/flate implementation
isn't ideal in terms of performance, but it's more than sufficient for
our needs, and it puts us on-par with rsync on amd64. We'll most likely
get some boost as codegen for other architectures improves (right now
we're at about 1.7x rsync transfer time for arm). I think the reason for
the boost is that by moving the FLATE compression to be internal, we're
at a better layer in the connection to benefit from compression.