Implemented transition tests for file swap and case check.
This commit implements two rather lengthy tests for transition: file
swap and checking for invalid path case on creation. Unfortunately these
can't be implemented more easily, or at least I can't think of a clever
way to do it. They are important tests though, so they are worth the
overhead.