Modified remote endpoint to strip executabililty for snapshot base.
In cases where a remote endpoint doesn't support executability, we were
still using the nominal ancestor as the rsync transfer base. This would
be inefficient in cases where the ancestor contained executability bits.
To get around this, we now strip executability from the ancestor before
using it as a base in the case where the remote endpoint doesn't
preserve executability.

We've also added an optimization to avoid propagating executability (and
therefore making a copy of the snapshot) if ancestor is nil, so we
should get a little bit of a benefit on new sessions where many files
are scanned.