Major refactor of session configuration.
This commit removes the dual-configuration session design in favor of a
single session configuration that is computed at creation time. It
seemed like the hypothetical "mutagen reload" design was going to be too
conceptually difficult to implement (even if one could understand it, it
would still be difficult to figure out what exactly its effects would
be), and thus it didn't make sense to keep this dual-configuration
design. Ignores have instead been kept separate in order to allow for
their reloading (which is potentially sensible) and better printing.

A VCS ignore option has also been added, with the default behavior being
to ignore VCS directories.