Modified PropagateExecutability to include ancestor.
Our scheme of propagating executability from the executability
preserving side to the non-preserving side was not entirely correct. A
problem would arise if the preserving side deleted a node with
executable files underneath it, thus not allowing propagation to the
equivalent node on the non-preserving side, showing a mismatch with
ancestor, and then repropagating the files from the non-preserving side
back to the preserving side (which showed deletion-only changes).

This commit modifes propagation to first try propagating from the
preserving side and then to try propagating from the ancestor. This
covers this edge case while still keeping the same behavior everywhere
else.