Disabled testing of Go tip on Travis CI.
We weren't getting much information from these reports (Go tip is
frequently broken) and supporting tip required some nasty version checks
in the Travis configuration file that had to be kept in sync.

In the future, when we need to switch to a version and generate a new
build, we can test it in a pull request first.