Fixed issues with transition tests.
It turns out that we had three bugs, one of which was being masked by
the other two.

First, we weren't providing the correct old entry when doing swap
transitions. But we weren't seeing this because (a) we were using
errors.Wrap around a nil error (so the modifier didn't appear to fail)
and we weren't correctly validating snapshots against expected when a
modifier was provided. It's fortunate we found these bugs. The
transition code was still correct, which is fortunate. It was only
through missing code coverage (that I expected to be present) that I
noticed these bugs.