Disabled FAT32 sync tests on Windows.
This commit disables use of the FAT32 partition for scan/transition
tests on Windows. It's not that we don't support FAT32 on Windows or
that it doesn't behave correctly, it's just that reparse points aren't
supported on FAT32, and thus symlinks fail to create, and we assume in
our scan/transition tests that we can freely create symlinks. In real
Mutagen usage, these will just register as failing to propagate and show
up like any other problem, but everything should still work.

We still use the FAT32 partition on Windows for other tests, and the
FAT32 implementations on macOS and other POSIX systems seem to support
symlinks (I guess via file metadata?) so we leave sync tests with FAT32
intact on those systems.