Added abililty to skip staging if files already staged.
We already had the ability to avoid staging files that were already
staged, but now we can jump out of staging entirely if all files are
already staged (probably rare, but theoretically possible). This is less
of an issue now that we've modified the rsync protocol to allow 0-file
staging, but it's still better to do to avoid spurious state callbacks.