Deleted superfluous integration tests.
We already perform mono-directional tests in both directions for local
endpoints, so there's no real need to do it in both directions over SSH
and in-memory SSH transports (we only really need to test
monodirectionally there).

This saves some time on integration tests, especially over these more
expensive remote transports. This is important because the macOS builder
keeps timing out (this test can take just over the 10 minutes of no
output that it fails on).