Disabled integration tests for macOS on Travis.
For reasons outlined in .travis.yml, this commit disables integration
tests on macOS on Travis. This isn't a huge deal, because integration
tests are still performed on Linux and Windows. The problem is basically
that it takes too long, and Travis times out builds that run for more
than 10 minutes without any output. Since macOS also has to do a release
build, it makes sense to not burden it with integration tests as well. I
think Travis' Mac hardware just isn't as vast or powerful as its Linux
hardware.