Empty commit for code coverage revival.
There's a super dangerous link on codecov.io that will instantly delete
the coverage history for a branch without any confirmation. It's also
disguised as a link, to make it even more of a footgun. This commit is
empty and will hopefully trigger codecov.io to start tracking master
again.