Increased wait time in watch tests.
This commit increases wait times between filesystem modifications in
watch tests. This seems to be necessary on the CI systems. In general,
native recursive filesystem watching is best-effort and
non-deterministic. If this doesn't fix the flake, we may need to just
skip recursive watching tests on CI systems. Everything seems to work
locally, so maybe it has something to do with FSEvents' behavior in a
virtualized environment.