Added auto-restart mechanism for recursive watches.
If native recursive filesystem watches throws an error or fails for
some reason, the native recursive watching function will now halt the
watch and attempt to recreate it. This will be useful in cases where,
e.g., the ReadDirectoryChangesW buffer overflows and is truncated due to
many events coming in at once. In these cases, the watch will simply be
recreated as opposed to erroring completely.