Reduced watch coalescing window.
The previous value of 100 ms was probably overly conservative. This
window only needs to be long enough to catch the next event if seeing a
series of rapid events (e.g. filesystem traversal/deletion), and 25 ms
is more than enough time for this. Using 25 ms will make things much
snappier.