Added fast path concatenation routine.
For Mutagen's internal path computations, we were using path.Join, but
this is unnecessarily slow for our simple use case (where we always know
that both componets will be in "clean" form). This shaves about 2% off
of scan time and probably a similar amount of overhead from
diff/reconcile/transition.