Revert "Switched to a far simpler and more robust watch root check."
This reverts commit 80bf73a393299fc508586ca74dba69190941a1a5.

It turns out that I was a bit daft and that the Windows implementation
here isn't actually correct, because os.SameFile loads file identifier
information from disk at the time of the comparison based on a stored
interal path, so it will always return true. I think it's more focused
on comparing hardlinks.