Removed kqueue support from hybrid watcher.
The kqueue implementation that we brought in from
github.com/rjeczalik/notify isn't quite right for our needs, so I've
removed it for the time being. The inotify implementation is a bit
simpler and easy to reason about, so I've left that for now because
hybrid watching is hugely valuable. We now have native watching of one
form or another on Windows, macOS, and Linux, which I think covers the
99% case.

Eventually we'll want to add support for kqueue (and potentially FEN),
but we're going to have to do custom implementations there.

I've refactored the hybrid watching code so that we can more easily add
separate non-native watch implementations in the future that don't
adhere to the notify package's Watcher interface. This brings the code
much closer in line with the recursive watching code.