Revert "Updated macOS recursive watcher to watch root parent."
This reverts commit 9a2eadd39293b08d7118d5fa6ae8ee4ad577325c.

This commit was problematic because FSEvents fully resolves any symlinks
in watch paths, and notifications come back with paths that are fully
resolved. This breaks our root parent check in cases when there's a
symlink in the watch path. The only way to fix this in a non-approximate
manner is to attempt to evaluate symlinks in the root path on a regular
basis (probably when we do the watch root check), but this adds a huge
amount of state complexity since EvalSymlinks can fail. The other option
would be matching base names, but that's expensive and approximate.