Modified FSEvents watcher to avoid watch recreation if unnecessary.
This commit adds an optimization to our recursive watcher to avoid
recreating watches on watch root changes if the root changes from file
to file. In this case, recreation is unnecessary. We conservatively
check that the mode stays the same, but this may be unnecessary, it
might suffice to only check if the type bits are 0 in both cases, but
this approach will handle any changes to readability or other permission
bits that might affect FSEvents now or in the future.