Added regular housekeeping to agent in endpoint mode.
This commit is a follow-up to the previous commit that added regular
housekeeping to the daemon. It adds regular housekeeping to the agent
when run in endpoint mode. This commit also fixes some less-than-ideal
behavior from the previous commit, such as not performing the initial
housekeeping until after the ticker fires for the first time. It also
restructures the code in a way that I think is a bit nicer.