Switched to more robust atomic operation filenames and added ignores.
This commit modifies WriteFileAtomic and RenameFileAtomic to use more
robust intermediate filenames. This is important because we rely on
rename failures to identify a number of cases where an invalid file base
name is being used, and re-using the base name as an intermediate file
name might make such errors less clear.

This commit also modifies watch and scan infrastructure to ignore these
intermediate files.