Renamed service packages to avoid conflicts.
Since we renamed the daemon service's Shutdown method to Terminate,
we've had a conflict with the session package. This was due to all of
the services living within a single virtual "service" package (in the
eyes of Protocol Buffers and gRPC). This commit instead changes the
names of the service packages to correspond to their parent packages.
This also required some service renaming because sometimes the parent
package contained a Protocol Buffers type with a conflicting name (e.g.
the Prompt and Session services had conflicting names). To solve this,
we renamed to Prompt and Session services to Prompting and Sessions,
which I think is a sane alternative (and perhaps a better name in the
case of Prompting). In the future, if the Go linker gets a bit better
and removing unused symbols, we may be able to fold these service
packages into their parent packages, and this commit helps to prepare
for that, but in the mean time we're unable to due to the code bloat
that comes with pulling in gRPC.