Delineated prompt and message streams.
This commit delineates prompting and messaging, allowing status messages
to be propagated back to create and resume commands. Eventually this
messaging will be expanded to pause and terminate commands.

Unfortunately this necessitated some pretty extensive changes, namely
making the prompt service a singleton. However, since this design is
actually pretty sane for us, and used extensively in the Go standard
library, I think we'll keep it. I've modified the daemon and sessions
services to follow this same pattern.