Refactored prompter registration and services.
This commit once again refactors prompter registration, switching to a
global service that is independent from the gRPC server. This removes
the introduction of a gRPC dependency on code performing messaging
(which had bloated agent binary size) and makes things significantly
cleaner in terms of implementation.

We've also moved away from the singleton service implementations, which
I think wasn't actually that good of a design.