Modified pause and terminate commands to support messaging.
This commit modifies the pause and terminate commands (and the
underlying RPC infrastructure) to support messages printed to the
console. Most of the time this isn't particularly useful, but sometimes
sessions take a little bit of time to fully halt, and now users will
have some feedback to help understand why.

Since we've had to implement most of the infrastructure for prompting on
pause and terminate streams, it might also make sense to allow for
remote clean-up on terminate streams by connecting to the remote, but
that's farther down the line.