Improved checks for alternate path separators.
This commit is a cleanup of the fix introduced in cca818e. Instead of
searching for "invalid characters", it now correctly notes that it is
only searching for alternate path separators (i.e. characters that have
a structural meaning to the specific operating system). There are only
two places that we need to check for this: at the root of a creation and
for individual entry names when creating a directory. For swaps and
removals, we already have an effective check for this with
ensureRouteWithProperCase. Thus, this new design is more elegant and
more efficient.