Added support for automatically relaunching inside winpty.
This commit modifies the mutagen command to automatically relaunch
inside of winpty if it detects a mintty console environment. Since this
functionality is now automatic, the corresponding documentation has been
removed.

Fortunately the package needed to support this was already available, so
there was no need to vendor anything new or add additional legal
information.