Changed endpoint stage method to take only entry digests.
This commit modifies endpoint's stage method to take a map from path to
digest as opposed to a map from path the entry. The digest is all that's
used in staging, meaning that all other information being transmitted
(especially on initial synchronizations) was wasteful, unnecessary, and
adding additional invariants that needed to be verified.