Updated build script to force running in source tree.
This commit updates the build script to enforce that its working
directory is inside the Mutagen source tree and to use a build directory
path that is always rooted at the source tree. This is necessary because
the build script itself assumes support for Go 1.11 modules.

This commit also disables the setting of GOPATH for commands run by the
build script. The primary reasons for this are (a) to allow a user to
wipe the build directory without wiping downloaded modules (making
re-builds significantly faster), (b) to avoid the difficulties of
removing the downloaded modules (which are all marked 0444), and (c) to
simplify the build script a bit.