Updated generate script for modules-based build.
This commit updates the generate script to work in a modules-based
environment. This actually vastly simplifies the generate script and
allows us to use much simpler imports.

The script still isn't at the point where it can be completely replaced
by a "go generate" command since we still need to build the generator,
but we're getting close.