Replaced generate script with "go generate" command.
It turns out that we can actually make our code generation work with
"go generate" now that we have module support to build the generator.

We don't really have a way to enforce that the user has module support
activated with this approach, but it will be the default soon, and we
can just make this requirement explicit in the instructions. Developers
are the only ones who should have a need to regenerate Protocol Buffers
files anyway.