Relocated service packages to conserve Protocol Buffers namespaces.
This commit consolidates the service packages so that we can keep their
package namespaces the same for Protocol Buffers while incorporating the
fixes needed for Go modules and go generate.

We may end up merging the service packages into their respective
packages, especially if we're going to bring a dependency on gRPC into
the agent binaries, at which point it will no longer be helpful to keep
service packages separate.