Added an additional heuristic to executability propagation.
This commit adds an additional heuristic to executability propagation
that propagates executability bits from the preserving side to the
non-preserving side in the case that the preserving side is unmodified
and the non-preserving side is modified. This is useful for editing
executable scripts on the non-preserving side where you want to maintain
executability on the preserving side (e.g. editing a script on Windows
that you're pushing over to Linux).

This is definitely a heuristic, because there are cases where you
wouldn't want this to be done, but these cases are vastly outnumbered by
cases where you would want such propagation. Additionally, it's far
easier to strip an executability bit once than it is to re-add it after
every file modification.