Massive refactoring in preparation for additional protocol support.
This commit extracts endpoint implementations to exist outside of the
session package, either in the local package or the agent package. This
is a much more natural organization and will make it much simpler to add
support for additional agent-based protocols or, eventually, gRPC-based
endpoints. The protocol handler registration system will also make it
possible for us to add support for gRPC-based endpoints without having
to resort to complex build tags to avoid bloating agent binaries.

This commit also updates endpoints to use the new rsync message types
and adds an optimization tweak to the rsync decoder API.