Fixed tests for tildeExpand lookups on Windows.
On Windows, os/user.Current returns a user object with a username that
has a format like DOMAIN\user. This commit adds a utility function for
tests to work around this. A similar fix isn't necessary inside
tildeExpand since the lookup already has the correct username and is
only trying to extract the home directory.