Performed some initial refactoring in preparation for custom protocols.
This commit performs some initial refactoring in preparation for custom
protocol support.

First, it relocates some definitions and agent package components that
will also be useful for custom protocol handling.

Second, it folds the handshake mechanism into the client/server
infrastructure, exposing enough information through error probing to
determine if the handshake failed due to a transport error. This will
also be useful for custom protocol handling, perhaps even more so than
for the agent package, so it makes sense to share the same
implementation.

Finally, it adds a configuration infrastructure to remote.ServeEndpoint
and local.NewEndpoint that allows for endpoint configuration validation
as well as overrides of synchronization root, cache path, staging root
path, and filesystem watch mechanism.