Implemented Docker protocol support.
This commit provides an initial implementation of Docker protocol
support. It currently lacks integration tests and hasn't been validated
on Windows. There are still a few TODOs in the agent and docker packages
for work that needs to be done.

Because of how Docker forwards standard error streams to standard
output, we've had to add a "magic number" feature to the remote endpoint
protocol.

We've also relocated some code that is useful across multiple
agent-based transport implementations.