Continued work on support for Docker Windows containers.
This commit continues work on support for Docker Windows containers. It
has a few rough edges due to Hyper-V limitations, but otherwise works
pretty well.

This change necessitated a re-think of our error classification
mechanism in agent dialing. This has now been delegated to the
individual transports, who are in a much better position to handle this.