Modified Docker URL format and fixed SSH URL parsing.
This commit modifies the Docker URL format to be closer to an
RFC-compliant URL. It also adds some fixes for SCP-style SSH URL parsing
to ensure that only valid URLs can be parsed out (it was previously
possible to end up with empty paths) and to catch other common errors.
This necessitated changing the port parsing semantics a bit, but only in
rare corner cases that I don't think will come up.