Disabled use of force poll watch mode in integration tests.
Now that we're no longer using the notify package on Windows and macOS,
I think we can switch back to native recursive watching. This may be the
culprit behind the poor performance of integration tests in the CI
environments, or at least a small part of it (I suspect that the bigger
issue is the lack of write buffer space causing many context switches on
systems with only 1 or 2 cores).