Disabled race detection for integration tests.
This is what was slowing down our integration tests by an order of
magnitude. We now do two rounds of tests for the native architecture:
one without race detection and one with race detection. For the latter,
we disable integration tests. Since this should speed things up
significantly, we've re-enabled it on macOS.

We've also restructured the configuration files slightly to accomodate
this and remove verbose output from the second and third test cycles.