Modified integration tests to run with race detection.
This commit re-enables end-to-end integration tests with the race
detector running and just uses a smaller synchronization root in these
cases. Tests without race detection continue to run with the full
$GOROOT/src directory for stress testing. It's important that the
end-to-end tests run with race detection since that's where we're
actually going to see race conditions.

To up the ante, we've also modified the end-to-end tests to run in
parallel, so we should really be able to suss out any race conditions.