Disabled parallel package test evaluation on CI system.
This commit disables parallel evaluation of packages in testing. While a
single package's tests run serially by default, separate packages are
processed in parallel. This becomes a problem because we have multiple
packages with tests that try to hold the daemon lock. This issue hadn't
cropped up yet on Travis and AppVeyor due to their limited number of
CPUs, but it could have. It definitely did locally.