Added tag component to version and set to "dev".
This commit adds a "tag" component to the version string that can be
used to identify a version as being special in some way (e.g. "dev" or
"beta1") while retaining the same version number. Its only effect at the
moment is to cause the agent binary to be stored in a different location
based on the tag. This will be useful to allow people to try beta
versions without needing to manually delete agents if they happen to
change before the final release.

For its initial value, I've set it to "dev".