Refactored service validation methods.
This commit extracts and extends our gRPC services' message validation
logic to match that of our internal protocol. Validation methods are now
attached to the messages themselves and are more widely used. This is
probably overkill for the mutual trust of localhost, but it will let us
watch for bugs in Protocol Buffers and gRPC, as well as broken internal
invariants.