Relocated signal handling initialization in daemon and endpoint.
This commit relocates signal handling initialization to occur before the
setup of servers and services in the daemon and endpoint entry points.
This allows us to ensure that the process isn't terminated by a signal
during initialization (which is just as bad as an unhandled termination
at any later point).