Added better early return for executability propagation.
When performing executability propgataion, we can also return early if
there's no source from which to propagate executability propagation.
This commit extends an existing early return check to watch for this
case. This should be particularly beneficial when a large file tree is
created on an endpoint which does not preserve executability
information, though it will only enhance performance on the initial
propagation operation.