Expanded reconciliation tests to include new conflict resolution modes.
This commit expands the existing reconciliation tests to include the new
conflict resolution modes. Many of the tests don't generate conflicts,
so they can be re-used. The tests which do generate conflicts have been
copied and tweaked for the relevant conflict resolution mode that they
exercise. To make this a bit easier, reconcileTestCase has been modified
to be able to run for multiple conflict resolution modes.