Added shorthand flag for synchronization mode.
This commit adds a shorthand flag (-m) to the create command for the
specification of synchronization mode. The thinking behind this is that
(a) this is the session mode that is most likely to be commonly switched
and (b) the word "synchronization" is long, tricky to type, and not even
consistently spelt between dialects of English.