Added permission exposure levels.
This commit extends the current (implicit) "portable" permission
propagation mode with a set of "permission exposure levels". This was
chosen (instead of explicit POSIX-inspired octal permission codes) in
order to maintain a portable abstraction that'll work across different
platform and to avoid the complexity associated with explicitly
specifying octal mode values that would be tricky to parse, difficult to
validate, and (most importantly) run up against algorithmic difficulties
when dealing with executability propagation.

These levels can currently be specified on a general basis (in which
case they'll apply to both endpoints), and endpoint-specific basis, or a
combination thereof.