Changed permission specification system and updated transition code.
This commit changes the permission specification mechanism (removing the
permission exposure level concept) and updates the transition
infrastructure to take advantage of the new race-free filesystem
infrastructure. Unfortunately this had to be done in the same commit
because the permission behavior in the transition code was too difficult
to shim. Tests are passing, though the code still needs an additional
read-through.