Implemented synchronization root sandboxing.
This commit uses the new race-free filesystem traversal infrastructure
to sandbox access to sychronization roots. Paths passed from a remote
can no longer escape the synchronization root via incorrect formatting
or a symbolic link.

This change affected both the endpoint implementation and rsync
implementation. In order to facilitate this, an Opener type was added to
the filesystem package, which provides race-free filesystem traversal
that is "fast" if paths are accessed in DFS traversal order, which
fortunately they are in our case.