Changed public-facing names for synchronization parameters.
While we'll use the full term "synchronization" within the Mutagen
codebase, this word is too cumbersome to type and not even consistently
spelled between different dialects of English. Instead, we'll use "sync"
as the short form of this word in all user-facing components.

This is part of a more general strategy we're taking, where the internal
form of a word will be fully spelled out, but user-facing inputs will be
shortened. E.g. "symbolic link" will be "symlink", "synchronization"
will be "sync", "maximum" will be "max", "minimum" will be "min", etc.

Errors will continue to use the full form of a word for clarity.