Implemented "no-watch" mode and flush command.
This commit adds a "no-watch" watching mode that makes filesystem
watching a no-op. Synchronization still occurs, but only when forced,
otherwise it just sits idle.

To that end, a "flush" command has also been added that forces a
synchronization cycle to occur for one or more sessions. This can be
used in conjunction with normal watching as well.