Modified controller to store and return slim conflicts.
This commit adds infrastructure to generate "slim" conflicts that don't
retain the entire entry tree structure underneath them. These conflicts
are still valid and return all of the information that we present to the
user already.

As part of this commit, the CopyShallow method of Entry has been renamed
to copySlim (and hence unexported). It wasn't needed outside of the sync
package, and the "shallow" moniker was misleading since it wasn't a
traditional "shallow" copy (i.e. one that preserves structure but only
duplicates top-level data).

This is part 1 of a 2-part fix for issue #51.