Modified controller to disable polling for endpoints with no watch.
This commit modifies the session controller to disable polling for
endpoints with a no-watch watch mode. In theory, these endpoints
shouldn't send back watch events, but this commit enforces that so that
a malfunctioning (or somehow malicious) endpoint can't force a
synchronization cycle.

This takes priority over endpoints with custom watch implementations,
making the controller the final authority on when synchronization
occurs.