Bumped Travis CI to Go version 1.11.9.
This commit bumps the Go version used on Travis CI (for testing and
builds) to Go 1.11.9. Technically this isn't essential, since it only
includes a fix for cgo on Linux (which we don't use), but I'd like to
keep the v0.8.x release series in step with the Go 1.11.x release
series.