Updated CacheEntry.Mode documentation.
This field no longer stores os.FileMode values on POSIX systems, instead
opting for the st_mode member of struct stat.

Now that we've changed the definition of this field's definition once
and realized that there are no consequences to doing so (other than a
full rescan), it also doesn't seem like the long-winded justification
for this field's definition was needed.