Implemented parallelized Directory.ReadContents.
This commit adds parallelization to Directory.ReadContents, using
runtime.NumCPU workers to perform parallel fstatat operations. This sees
a significant (~30%) drop in scan time.

The parallelization is implemented using a single global Goroutine pool.
This has the benefit of regulating disk contention in a daemon handling
many sessions simultaneously. If we were to allow the instantiation of
multiple pools (by encapsulating their implementation in a type), we'd
likely do more harm than good. As it is, there should be almost no
contention for the global pool, and if there is, then it would be made
worse by distributing work across more pools.

The one potential issue with this approach is that fstatat, which
underlies these metadata queries, does not guarantee that it is safe to
call concurrently on the same underlying directory file descriptor,
though in practice it seems safe to do so on most platforms.

Since this implementation is experimental, and environment variable
(MUTAGEN_PARALLEL_METADATA_DISABLED) has been added to switch back to
the serial implementation.